Add cross-fetch integration test, and test integrations separately
