+++ package.json
@@ -15,7 +15,7 @@
   ],
   "main": "index.js",
   "scripts": {
-    "lint": "eslint index.js test.js",
+    "lint": "eslint index.js test",
     "start": "cd demo && yarn start",
     "test": "yarn run lint && yarn test:coverage",
     "test:coverage": "jest --coverage",
@@ -31,6 +31,7 @@
   },
   "devDependencies": {
     "coveralls": "^3.0.4",
+    "cross-fetch": "^3.0.4",
     "eslint": "^5.16.0",
     "isomorphic-fetch": "^2.2.1",
     "isomorphic-unfetch": "^3.0.0",
