+++ test/cross-fetch.test.js
@@ -1,0 +1,6 @@
+const baseFetch = require("cross-fetch");
+const defineTests = require("./tests");
+
+describe("using cross-fetch", () => {
+  defineTests(baseFetch);
+});
