+++ test/isomorphic-fetch.test.js
@@ -1,0 +1,6 @@
+const baseFetch = require("isomorphic-fetch");
+const defineTests = require("./tests");
+
+describe("using isomorphic-fetch", () => {
+  defineTests(baseFetch);
+});
