+++ test/isomorphic-unfetch.test.js
@@ -1,0 +1,6 @@
+const baseFetch = require("isomorphic-unfetch");
+const defineTests = require("./tests");
+
+describe("using isomorphic-unfetch", () => {
+  defineTests(baseFetch);
+});
