+++ test/node-fetch.test.js
@@ -1,0 +1,6 @@
+const baseFetch = require("node-fetch");
+const defineTests = require("./tests");
+
+describe("using node-fetch", () => {
+  defineTests(baseFetch);
+});
